seven

+++ include/monitor.h
@@ -2,7 +2,7 @@
 #define MONITOR_H
 
 // 监控模块获取的值
-extern float MonitorValue[4];
+extern int MonitorValue[4];
 
 // 引脚初始化
 void MonitorPinInit();
